Actualizados LogIn, LogUp, RecoverPassword y SignOut con Firebase Auth
