+++ .gitignore
@@ -83,3 +83,6 @@ dist
 # Statics files
 *.mp4
 *.avi
+
+#Firebase configuration
+fbconfig.js
