+++ package.json
@@ -1,12 +1,8 @@
 {
-  "name": "postic",
+  "name": "postic-master-firebase",
   "version": "2.0.0",
   "description": "App to Point Of Sale. Final student project Escuela JS 2019 Platzi.",
-  "keywords": [
-    "Point Of Sale App",
-    "ReactJS",
-    "Create React APP"
-  ],
+  "keywords": [],
   "main": "src/index.js",
   "author": "Caraballo Luis, Fajardo Mauricio, Nemogá Jorge",
   "license": "MIT",
@@ -33,6 +29,7 @@
     "@material-ui/icons": "4.4.3",
     "@material-ui/styles": "^4.6.0",
     "clsx": "1.0.4",
+    "firebase": "7.4.0",
     "history": "^4.10.1",
     "material-table": "1.52.0",
     "prop-types": "^15.7.2",
