--- postic_with_firebase/README.md
@@ -1,30 +1,0 @@
-# POSTIC: Proyecto final Escuela JS - Platzi 2019
-
-## Datos del grupo
-
-**Temática**: Sistema Punto de Venta (POS - Point Of Sale)
-
-**Integrantes**:
-
-- Jorge Nemogá
-- Luis Caraballo
-- Mauricio Fajardo
-
-## Sistema _POS_
-
-Un sistema POS o Point Of Sale en este contexto hace referencia a aplicaciones y software destinado a la sistematización del control de ventas e inventarios de un negocio. El objetivo de este sistema es mantener el inventario actualizado de acuerdo con las ventas e ingreso de productos al negocio, optimizar el proceso de facturación para agilizar y mejorar los procesos de negocio, llevar el control de gastos e ingresos del negocio.
-
-## Entorno de desarrollo
-
-El archivo `.nvmrc` indica la versión de node.js a usar en el proyecto, se estará usando la última versión para el momento **node.js 10.16.3**.
-
-La documentación inicial se presentará en la [wiki del repositorio](#).
-
-## Vistas
-
-- Login
-- SignUp
-- Nueva Venta
-- Informes
-- Pedidos
-- Productos
