--- postic_with_firebase/package.json
@@ -1,81 +1,0 @@
-{
-  "name": "postic",
-  "version": "2.0.0",
-  "description": "App to Point Of Sale. Final student project Escuela JS 2019 Platzi.",
-  "keywords": [
-    "Point Of Sale App",
-    "ReactJS",
-    "Create React APP"
-  ],
-  "main": "src/index.js",
-  "author": "Caraballo Luis, Fajardo Mauricio, Nemogá Jorge",
-  "license": "MIT",
-  "homepage": "https://github.com/MauricioFa/POSTIC#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/MauricioFa/POSTIC.git"
-  },
-  "bugs": {
-    "url": "https://github.com/MauricioFa/POSTIC/issues"
-  },
-  "scripts": {
-    "build": "webpack --mode production",
-    "start": "webpack-dev-server --open --mode development",
-    "format": "prettier --write '{*.js,src/**/*.{js,jsx}}'",
-    "linter": "eslint src/ --fix",
-    "linterNoFix": "eslint src/"
-  },
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-brands-svg-icons": "^5.11.2",
-    "@fortawesome/react-fontawesome": "^0.1.7",
-    "@material-ui/core": "4.5.1",
-    "@material-ui/icons": "4.4.3",
-    "@material-ui/styles": "^4.6.0",
-    "clsx": "1.0.4",
-    "firebase": "7.4.0",
-    "history": "^4.10.1",
-    "material-table": "1.52.0",
-    "prop-types": "^15.7.2",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-redux": "^7.1.1",
-    "react-router": "^5.1.2",
-    "react-router-dom": "5.1.2",
-    "recharts": "1.7.1",
-    "redux": "^4.0.4",
-    "webpack": "4.41.0",
-    "webpack-cli": "^3.3.10"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.6.4",
-    "@babel/preset-env": "^7.6.3",
-    "@babel/preset-react": "^7.6.3",
-    "@svgr/webpack": "^4.3.3",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^3.2.0",
-    "eslint": "^6.5.1",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.4.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
-    "file-loader": "^4.2.0",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.12.0",
-    "prettier": "^1.18.2",
-    "sass": "^1.23.3",
-    "sass-loader": "^8.0.0",
-    "webpack-dev-server": "^3.9.0"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
-}
