--- postic_with_firebase/src/assets/styles/ShoppingCart.css
@@ -1,29 +1,0 @@
-.Checkout {
-    grid-template-columns: 3fr 1fr;
-    grid-gap: 2rem;
-    display: grid ;
-}
-
-.Checkout-item {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.Checkout-item .fas {
-    margin: 0 0 0 10px;
-    color:  rgba(0,0,0,.4);
-    cursor: pointer;
-}
-
-.Checkout-element {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    border-bottom: 1px solid #eee;
-}
-
-.Checkout-element h4 {
-    margin: 0;
-}
