--- postic_with_firebase/src/assets/styles/styles.css
@@ -1,13 +1,0 @@
-*, *::before, *::after, html {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  margin-block-start: 0em;
-  margin-block-end: 0em;
-  margin-inline-start: 0px;
-  margin-inline-end: 0px;
-}
-
-body {
-  font-size: 16px;
-}
