--- postic_with_firebase/src/containers/Dashboard.jsx
@@ -1,67 +1,0 @@
-import React from 'react';
-import clsx from 'clsx';
-import { makeStyles } from '@material-ui/core/styles';
-import { Container, Grid, Paper } from '@material-ui/core';
-import AddNewButton from '../components/AddNewButton';
-import Chart from '../components/Chart';
-import CopyrightLabel from '../components/CopyrightLabel';
-import Deposits from '../components/Deposits';
-import ListRecentOrders from '../components/ListRecentOrders';
-
-const useStyles = makeStyles((theme) => ({
-  appBarSpacer: theme.mixins.toolbar,
-  content: {
-    flexGrow: 1,
-    height: '100vh',
-    overflow: 'auto',
-  },
-  container: {
-    paddingTop: theme.spacing(4),
-    paddingBottom: theme.spacing(4),
-  },
-  paper: {
-    padding: theme.spacing(2),
-    display: 'flex',
-    overflow: 'auto',
-    flexDirection: 'column',
-  },
-  fixedHeight: {
-    height: 340,
-  },
-}));
-
-const Dashboard = () => {
-  const classes = useStyles();
-  const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
-
-  return (
-    <>
-      <main className={classes.content}>
-        <div className={classes.appBarSpacer} />
-        <Container maxWidth='lg' className={classes.container}>
-          <Grid container spacing={3}>
-            <Grid item xs={12} md={8} lg={9}>
-              <Paper className={fixedHeightPaper}>
-                <Chart />
-              </Paper>
-            </Grid>
-            <Grid item xs={12} md={4} lg={3}>
-              <Paper className={fixedHeightPaper}>
-                <Deposits />
-              </Paper>
-            </Grid>
-            <Grid item xs={12}>
-              <Paper className={classes.paper}>
-                <ListRecentOrders />
-              </Paper>
-            </Grid>
-          </Grid>
-        </Container>
-        <CopyrightLabel title='Copyright' name='POSTIC' date='2019' linkTo='/#' />
-        <AddNewButton />
-      </main>
-    </>
-  );
-};
-
-export default Dashboard;
