--- postic_with_firebase/src/containers/RegisterSale.jsx
@@ -1,59 +1,0 @@
-import React from 'react';
-import clsx from 'clsx';
-import { makeStyles } from '@material-ui/core/styles';
-import { Container, Grid, Paper } from '@material-ui/core';
-import CopyrightLabel from '../components/CopyrightLabel';
-import Inventory from '../components/Inventory';
-import ShoppingCart from '../components/ShoppingCart';
-
-const useStyles = makeStyles((theme) => ({
-  appBarSpacer: theme.mixins.toolbar,
-  content: {
-    flexGrow: 1,
-    height: '100vh',
-    overflow: 'auto',
-  },
-  container: {
-    paddingTop: theme.spacing(4),
-    paddingBottom: theme.spacing(4),
-  },
-  paper: {
-    padding: theme.spacing(2),
-    display: 'flex',
-    overflow: 'auto',
-    flexDirection: 'column',
-  },
-  fixedHeight: {
-    height: 640,
-  },
-}));
-
-const RegisterSale = () => {
-  const classes = useStyles();
-  const fixedHeightPaper = clsx(classes.paper, classes.fixedHeight);
-
-  return (
-    <>
-      <main className={classes.content}>
-        <div className={classes.appBarSpacer} />
-        <Container maxWidth='lg' className={classes.container}>
-          <Grid container spacing={3}>
-            <Grid item xs={12} md={8} lg={8}>
-              <Paper className={fixedHeightPaper}>
-                <Inventory />
-              </Paper>
-            </Grid>
-            <Grid item xs={12} md={4} lg={4}>
-              <Paper className={fixedHeightPaper}>
-                <ShoppingCart />
-              </Paper>
-            </Grid>
-          </Grid>
-        </Container>
-        <CopyrightLabel title='Copyright' name='POSTIC' date='2019' linkTo='/#' />
-      </main>
-    </>
-  );
-};
-
-export default RegisterSale;
