--- postic_with_firebase/src/index.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { createStore } from 'redux';
-import { Provider } from 'react-redux';
-import { Router } from 'react-router';
-import { createBrowserHistory } from 'history';
-import App from './routes/App';
-import reducers from './reducers/indexReducers';
-import initialState from './mocks/initialState';
-
-const store = createStore(reducers, initialState);
-const history = createBrowserHistory();
-
-ReactDOM.render(
-  <Provider store={store}>
-    <Router history={history}>
-      <App />
-    </Router>
-  </Provider>,
-  document.getElementById('root')
-);
